# server/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: OpenAPI Invoice generator specification
  version: 0.1.0
servers:
  - url: http://localhost:5000/
    description: Dev server
paths:
  /:
    get:
      summary: method for generating and mailing an invoice
      tags:
        - Invoice
      operationId: getInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addWork'
      responses:
        '200':
          description: Result of Sent Invoice
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/email'
        default:
          description: anything out of the ordinary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: method for adding a task list
      tags:
        - Invoice
      operationId: createListWorks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/listWorks'
      responses:
        '200':
          description: Result of Sent Invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invoice'
        default:
          description: anything out of the ordinary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    sentInvoice:
      type: string
      example: Succes Sent
    email:
      type: string
      example: example@gmail.com
    addWork:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: example@gmail.com
    listWorks:
      type: object
      required:
        - email
        - works
      properties:
        email:
          type: string
          example: example@gmail.com
        works:
          type: array
          items:
            $ref: '#/components/schemas/work'
    work:
      type: object
      required:
        - nameWork
        - priceWork
      properties:
        nameWork:
          type: string
          example: make design
        priceWork:
          type: number
          example: 100.12
    invoice:
      type: object
      required:
        - email
        - id
        - firstName
        - lastName
        - company
        - createdAt
        - updatedAt
      properties:
        email:
          $ref: '#/components/schemas/email'
        id:
          type: integer
          example: 1
        firstName:
          type: string
          example: Bob
        lastName:
          type: string
          example: Hill
        company:
          type: string
          example: Google
        createdAt:
          type: string
          example: 2023-04-10T10:35:53.431Z
        updatedAt:
          type: string
          example: 2023-04-10T10:35:53.431Z
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string